+++ rules/consistent-compose.js
@@ -1,6 +1,10 @@
 'use strict';
 
+var _ = require('lodash/fp');
 var enhance = require('./core/enhance');
+
+var knownComposeMethods = ['flow', 'flowRight', 'compose', 'pipe'];
+var isKnownComposeMethod = _.includes(_, knownComposeMethods);
 
 module.exports = function (context) {
   var info = enhance();
@@ -12,7 +16,7 @@ module.exports = function (context) {
   return info.merge({
     CallExpression: function (node) {
       var name = info.helpers.isMethodCall(node);
-      if (name !== false && name !== composeMethod) {
+      if (isKnownComposeMethod(name) && name !== composeMethod) {
         context.report(node, 'Forbidden use of `' + name + '`. Use `' + composeMethod + '` instead');
       }
     }
@@ -21,5 +25,5 @@ module.exports = function (context) {
 
 module.exports.schema = [{
   type: 'string',
-  enum: ['flow', 'flowRight', 'compose', 'pipe']
+  enum: knownComposeMethods
 }];